+++ app/assets/javascripts/bootsy/modal.js
@@ -42,14 +42,47 @@ Bootsy.Modal = function(area) {
     this.deleteImage(data.id);
   }.bind(this));
 
-  this.$el.on('ajax:error', '.bootsy-upload-form', this.imageUploadFailed.bind(this));
-
   this.$el.on('click', 'a[href="#refresh-gallery"]', this.requestImageGallery.bind(this));
 
-  this.$el.on('ajax:success', '.bootsy-upload-form', function(_e, data) {
-    this.area.setImageGalleryId(data.gallery_id);
-    this.addImage(data.image);
-    this.setUploadForm(data.form);
+  this.$el.on('submit', '.bootsy-upload-form', function(event, xhr, settings) {
+    var fileSelect = event.target.querySelector('input[type="file"]');
+    var formData = new FormData();
+    var file = fileSelect.files[0];
+    var fileURLInputName = 'image[remote_image_file_url]';
+    var fileURLInput = event.target.querySelector(
+      'input[name="' + fileURLInputName + '"]');
+    var fileURL;
+
+    event.preventDefault();
+
+    formData.append('authenticity_token',
+      event.target.querySelector('input[name="authenticity_token"]').value);
+
+    if (file) {
+      formData.append('image[image_file]', file, file.name);
+    }
+
+    if (fileURLInput) {
+      fileURL = fileURLInput.value;
+    } else {
+      fileURL = '';
+    }
+
+    formData.append(fileURLInputName, fileURL);
+
+    var xhr = new XMLHttpRequest();
+    xhr.open('POST', event.target.action, true);
+    xhr.onload = function () {
+      var data = JSON.parse(xhr.response);
+      if (xhr.status === 200) {
+        this.area.setImageGalleryId(data.gallery_id);
+        this.addImage(data.image);
+        this.setUploadForm(data.form);
+      } else {
+        this.imageUploadFailed(xhr, data);
+      }
+    }.bind(this);
+    xhr.send(formData);
   }.bind(this));
 
   this.$el.modal({ show: false });
@@ -122,9 +155,7 @@ Bootsy.Modal.prototype.setUploadForm = function(html) {
 };
 
 // The image upload failed
-Bootsy.Modal.prototype.imageUploadFailed = function(_e, xhr, _status, error) {
-  var invalidErrors = xhr.responseJSON;
-
+Bootsy.Modal.prototype.imageUploadFailed = function(xhr, invalidErrors) {
   if (Number(xhr.status) === 422 && invalidErrors.image_file) {
     this.hideUploadLoadingAnimation();